created separate file for getIsOverflowHidden
