+++ src/extensions/styles/getIsOverflowHidden.js
@@ -1,0 +1,19 @@
+/**
+ * Internal dependencies
+ */
+
+import getLastBreakpointAttribute from './getLastBreakpointAttribute';
+
+export const getIsOverflowHidden = (attributes, breakpoint) =>
+	getLastBreakpointAttribute({
+		target: 'overflow-y',
+		breakpoint,
+		attributes,
+	}) === 'hidden' &&
+	getLastBreakpointAttribute({
+		target: 'overflow-x',
+		breakpoint,
+		attributes,
+	}) === 'hidden';
+
+export default getIsOverflowHidden;
