+++ src/extensions/styles/helpers/getClipPathStyles.js
@@ -1,4 +1,4 @@
-import { getLastBreakpointAttribute } from '../getLastBreakpointAttribute';
+import getLastBreakpointAttribute from '../getLastBreakpointAttribute';
 /**
  * General
  */
