+++ src/extensions/styles/helpers/getColumnSizeStyles.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { getLastBreakpointAttribute } from '../getLastBreakpointAttribute';
+import getLastBreakpointAttribute from '../getLastBreakpointAttribute';
 
 /**
  * External dependencies
