+++ src/extensions/styles/helpers/getOverflowStyles.js
@@ -1,4 +1,4 @@
-import { getLastBreakpointAttribute } from '../getLastBreakpointAttribute';
+import getLastBreakpointAttribute from '../getLastBreakpointAttribute';
 
 /**
  * General
