+++ src/extensions/styles/helpers/getSVGStyles.js
@@ -3,7 +3,7 @@
  * Internal dependencies
  */
 import getColorRGBAString from '../getColorRGBAString';
-import { getLastBreakpointAttribute } from '../getLastBreakpointAttribute';
+import getLastBreakpointAttribute from '../getLastBreakpointAttribute';
 import getPaletteAttributes from '../getPaletteAttributes';
 
 /**
