+++ src/extensions/styles/helpers/getSizeStyles.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { getLastBreakpointAttribute } from '../getLastBreakpointAttribute';
+import getLastBreakpointAttribute from '../getLastBreakpointAttribute';
 import getDefaultAttribute from '../getDefaultAttribute';
 
 /**
