+++ src/extensions/styles/index.js
@@ -10,8 +10,8 @@ export { default as getColorRGBAString } from './getColorRGBAString';
 export { default as getDefaultAttribute } from './getDefaultAttribute';
 export { default as getGroupAttributes } from './getGroupAttributes';
 export { default as getIconWithColor } from './getIconWithColor';
-export { getLastBreakpointAttribute } from './getLastBreakpointAttribute';
-export { getIsOverflowHidden } from './getLastBreakpointAttribute';
+export { default as getLastBreakpointAttribute } from './getLastBreakpointAttribute';
+export { default as getIsOverflowHidden } from './getIsOverflowHidden';
 export { default as getPaletteAttributes } from './getPaletteAttributes';
 export { default as hoverAttributesCreator } from './hoverAttributesCreator';
 export { default as paletteAttributesCreator } from './paletteAttributesCreator';
